hldjango/requirements.txt
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
backports-tarfile==1.2.0 ; python_version >= "3.10" and python_version < "3.12"
black==24.10.0 ; python_version >= "3.10" and python_version < "4.0"
build==1.2.2.post1 ; python_version >= "3.10" and python_version < "4.0"
cachecontrol==0.14.2 ; python_version >= "3.10" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy" or python_version >= "3.10" and python_version < "4.0" and sys_platform == "darwin"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows" or python_version >= "3.10" and python_version < "4.0" and os_name == "nt"
crashtest==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
crispy-bootstrap4==2024.10 ; python_version >= "3.10" and python_version < "4.0"
crispy-bootstrap5==2024.10 ; python_version >= "3.10" and python_version < "4.0"
cryptography==44.0.1 ; python_version >= "3.10" and python_version < "4.0"
distlib==0.3.9 ; python_version >= "3.10" and python_version < "4.0"
django-ace==1.37.5 ; python_version >= "3.10" and python_version < "4.0"
django-active-link==0.1.8 ; python_version >= "3.10" and python_version < "4.0"
django-allauth==0.62.1 ; python_version >= "3.10" and python_version < "4.0"
django-configurations==2.5.1 ; python_version >= "3.10" and python_version < "4.0"
django-crispy-forms==2.3 ; python_version >= "3.10" and python_version < "4.0"
django-dbbackup==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
django-debug-toolbar==5.0.1 ; python_version >= "3.10" and python_version < "4.0"
django-extensions==3.2.3 ; python_version >= "3.10" and python_version < "4.0"
django-ipware==7.0.1 ; python_version >= "3.10" and python_version < "4.0"
django-sendgrid-v5==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
django==5.1.6 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.2 ; python_version >= "3.10" and python_version < "4.0"
dulwich==0.22.7 ; python_version >= "3.10" and python_version < "4.0"
fastjsonschema==2.21.1 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.17.0 ; python_version >= "3.10" and python_version < "4.0"
graphviz==0.20.3 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.10" and python_version < "4.0"
huey==2.5.2 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version >= "3.10" and python_version < "3.12"
installer==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
jaraco-classes==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
jaraco-context==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
jaraco-functools==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
jeepney==0.8.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux"
keyring==25.6.0 ; python_version >= "3.10" and python_version < "4.0"
lark==1.2.2 ; python_version >= "3.10" and python_version < "4.0"
mistletoe==1.4.0 ; python_version >= "3.10" and python_version < "4.0"
more-itertools==10.6.0 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
ordered-set==4.1.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
paste==3.10.1 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
pdf2image==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
pillow==11.1.0 ; python_version >= "3.10" and python_version < "4.0"
pkginfo==1.12.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==4.3.6 ; python_version >= "3.10" and python_version < "4.0"
poetry-core==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
poetry-plugin-export==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
poetry==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
pycipher==0.5.2 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy" or python_version >= "3.10" and python_version < "4.0" and sys_platform == "darwin"
pydot==3.0.4 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
pylatex==1.4.2 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
pyproject-hooks==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
python-http-client==3.3.7 ; python_version >= "3.10" and python_version < "4.0"
python-ipware==3.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.10" and python_version < "4.0"
pywin32-ctypes==0.2.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
rapidfuzz==3.12.1 ; python_version >= "3.10" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux"
sendgrid==6.11.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==75.8.0 ; python_version >= "3.10" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.10" and python_version < "4.0"
starkbank-ecdsa==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.2.1 ; python_version >= "3.10" and python_version < "3.11"
tomlkit==0.13.2 ; python_version >= "3.10" and python_version < "4.0"
trove-classifiers==2025.1.15.22 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2025.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
verbose-csrf-middleware==1.0.4 ; python_version >= "3.10" and python_version < "4.0"
virtualenv==20.29.2 ; python_version >= "3.10" and python_version < "4.0"
waitress==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
whitenoise==6.9.0 ; python_version >= "3.10" and python_version < "4.0"
xattr==1.1.4 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "darwin"
zipp==3.21.0 ; python_version >= "3.10" and python_version < "3.12"
